refactor(stripe): migrate Stripe service methods to stripeClient for improved structure and remove deprecated third-party service file
