+++ web-app/src/app/(app)/billing/page.tsx
@@ -1,13 +1,14 @@
 import BillingForm from "@/components/billing/billing-form";
 import { getEnvSecrets } from "@/config/env.secrets";
 import { getSessionOrRedirect } from "@/lib/auth/utils";
-import { getActiveOrganization, getPriceFromProductId } from "@/lib/services";
+import { stripeClient } from "@/lib/clients/stripe.client";
+import { getActiveOrganization } from "@/lib/services";
 
 export default async function BillingPage() {
   await getSessionOrRedirect();
 
   const productId = getEnvSecrets().STRIPE_PRODUCT_ID;
-  const price = await getPriceFromProductId(productId);
+  const price = await stripeClient.getPriceByProductId(productId);
   const activeOrganization = await getActiveOrganization();
 
   return (
