+++ web-app/src/lib/services/stripe.service.ts
@@ -1,0 +1,1 @@
+export const stripeService = {};
